formatted the base template with line breaks. changed some input fields from simple text to the appropriate format. added jinja block to title. added title to list.html head, border, th tags, img links for edit/delete to the table. renamed story.html to base, create.html to story. pulled static dir from git ignore, and changed the tables est col. to float type from int

+++ models.py
@@ -13,7 +13,7 @@ class UserStoryManager(BaseModel):
     user_story = TextField()
     acceptance_criteria = TextField()
     business_value = IntegerField()
-    estimation = IntegerField()
+    estimation = FloatField()
     status = CharField()
